+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniApp.kt
@@ -34,34 +34,30 @@ abstract class MiniApp internal constructor() {
      * Creates a mini app.
      * The mini app is downloaded, saved and provides a [MiniAppDisplay] when successful.
      * @param appId mini app id.
-     * @param miniAppMessageBridge the interface for communicating between host app & mini app
-     * @throws MiniAppSdkException when there is some issue during fetching,
+     * @param miniAppMessageBridge the interface for communicating between host app & mini app.
+     * @param miniAppNavigator allow host app to handle specific urls such as external link.
+     * @throws [MiniAppNotFoundException] when the specified mini app ID does not exist on the server
+     * @throws [MiniAppHasNoPublishedVersionException] when the specified mini app ID exists on the
+     * server but has no published versions
+     * @throws [MiniAppSdkException] when there is any other issue during fetching,
      * downloading or creating the view.
      */
-    @Throws(MiniAppSdkException::class)
-    abstract suspend fun create(
-        appId: String,
-        miniAppMessageBridge: MiniAppMessageBridge
-    ): MiniAppDisplay
-
-    /**
-     * Same as [create(String, MiniAppMessageBridge)].
-     * Use this to control external url loader.
-     * @param miniAppNavigator allow host app to handle specific urls such as external link.
-     */
-    @Throws(MiniAppSdkException::class)
+    @Throws(MiniAppNotFoundException::class, MiniAppHasNoPublishedVersionException::class, MiniAppSdkException::class)
     abstract suspend fun create(
         appId: String,
         miniAppMessageBridge: MiniAppMessageBridge,
-        miniAppNavigator: MiniAppNavigator
+        miniAppNavigator: MiniAppNavigator? = null
     ): MiniAppDisplay
 
     /**
      * Fetches meta data information of a mini app.
      * @return [MiniAppInfo] for the provided appId of a mini app
-     * @throws [MiniAppSdkException] when fetching fails from the BE server for any reason.
+     * @throws [MiniAppNotFoundException] when the specified mini app ID does not exist on the server
+     * @throws [MiniAppHasNoPublishedVersionException] when the specified mini app ID exists on the
+     * server but has no published versions
+     * @throws [MiniAppSdkException] when fetching fails from the BE server for any other reason.
      */
-    @Throws(MiniAppSdkException::class)
+    @Throws(MiniAppNotFoundException::class, MiniAppHasNoPublishedVersionException::class, MiniAppSdkException::class)
     abstract suspend fun fetchInfo(appId: String): MiniAppInfo
 
     /**
@@ -115,10 +111,9 @@ abstract class MiniApp internal constructor() {
             defaultConfig = miniAppSdkConfig
             val apiClient = ApiClient(
                 baseUrl = miniAppSdkConfig.baseUrl,
-                rasAppId = miniAppSdkConfig.rasAppId,
+                rasProjectId = miniAppSdkConfig.rasProjectId,
                 subscriptionKey = miniAppSdkConfig.subscriptionKey,
-                hostAppVersionId = miniAppSdkConfig.hostAppVersionId,
-                isTestMode = miniAppSdkConfig.isTestMode
+                isPreviewMode = miniAppSdkConfig.isPreviewMode
             )
             val apiClientRepository = ApiClientRepository().apply {
                 registerApiClient(defaultConfig.key, apiClient)