+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniAppSdkConfig.kt
@@ -3,32 +3,49 @@ package com.rakuten.tech.mobile.miniapp
 /**
  * This represents the configuration settings for the Mini App SDK.
  * @property baseUrl Base URL used for retrieving a Mini App.
- * @property rasAppId App ID for the Platform API.
+ * @property rasProjectId Project ID for the Platform API.
  * @property subscriptionKey Subscription Key for the Platform API.
  * @property hostAppVersionId Version of the host app, used to determine feature compatibility for Mini App.
  * @property hostAppUserAgentInfo User Agent information from Host App.
- * @property isTestMode Whether the sdk is making use of the Test API Endpoints for under "Testing" mini apps.
+ * @property isPreviewMode Whether the host app wants to use the API Endpoints under "Preview" mode.
  */
 data class MiniAppSdkConfig(
     val baseUrl: String,
-    val rasAppId: String,
+    val rasProjectId: String,
     val subscriptionKey: String,
-    val hostAppVersionId: String,
+    val hostAppVersionId: String = "",
     val hostAppUserAgentInfo: String,
-    val isTestMode: Boolean
+    val isPreviewMode: Boolean
 ) {
-    internal val key = "$baseUrl-$isTestMode-$rasAppId-$subscriptionKey-$hostAppVersionId"
+    internal val key = "$baseUrl-$isPreviewMode-$rasProjectId-$subscriptionKey"
+
+    @Deprecated("Use isPreviewMode instead of isTestMode, Use rasProjectId instead of rasAppId")
+    constructor(
+        baseUrl: String,
+        rasAppId: String,
+        rasProjectId: String = rasAppId,
+        subscriptionKey: String,
+        hostAppVersionId: String = "",
+        hostAppUserAgentInfo: String,
+        isTestMode: Boolean,
+        isPreviewMode: Boolean = isTestMode
+    ) : this(
+        baseUrl = baseUrl,
+        rasProjectId = rasProjectId,
+        subscriptionKey = subscriptionKey,
+        hostAppVersionId = hostAppVersionId,
+        hostAppUserAgentInfo = hostAppUserAgentInfo,
+        isPreviewMode = isPreviewMode
+    )
 
     init {
         when {
             !isBaseUrlValid(baseUrl) ->
                 throw sdkExceptionForInvalidArguments("MiniAppSdkConfig with invalid baseUrl")
-            rasAppId.isBlank() ->
-                throw sdkExceptionForInvalidArguments("MiniAppSdkConfig with invalid rasAppId")
+            rasProjectId.isBlank() ->
+                throw sdkExceptionForInvalidArguments("MiniAppSdkConfig with invalid rasProjectId")
             subscriptionKey.isBlank() ->
                 throw sdkExceptionForInvalidArguments("MiniAppSdkConfig with invalid subscriptionKey")
-            hostAppVersionId.isBlank() ->
-                throw sdkExceptionForInvalidArguments("MiniAppSdkConfig with invalid hostAppVersionId")
         }
     }
 }