+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/MiniAppSdkConfigSpec.kt
@@ -9,24 +9,22 @@ class MiniAppSdkConfigSpec {
     fun `the key pattern should match the defined scheme`() {
         val config = MiniAppSdkConfig(
             baseUrl = TEST_URL_HTTPS_2,
-            isTestMode = true,
-            rasAppId = TEST_HA_ID_APP,
+            isPreviewMode = true,
+            rasProjectId = TEST_HA_ID_PROJECT,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
-            hostAppVersionId = TEST_HA_ID_VERSION,
             hostAppUserAgentInfo = TEST_HA_NAME
         )
-        config.key shouldEqual "${config.baseUrl}-${config.isTestMode}" +
-                "-${config.rasAppId}-${config.subscriptionKey}-${config.hostAppVersionId}"
+        config.key shouldEqual "${config.baseUrl}-${config.isPreviewMode}" +
+                "-${config.rasProjectId}-${config.subscriptionKey}"
     }
 
     @Test(expected = MiniAppSdkException::class)
     fun `should throw exception when api url is not https`() {
         MiniAppSdkConfig(
             baseUrl = "http://www.example.com/1",
-            isTestMode = false,
-            rasAppId = TEST_HA_ID_APP,
+            isPreviewMode = false,
+            rasProjectId = TEST_HA_ID_PROJECT,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
-            hostAppVersionId = TEST_HA_ID_VERSION,
             hostAppUserAgentInfo = TEST_HA_NAME
         )
     }
@@ -35,22 +33,20 @@ class MiniAppSdkConfigSpec {
     fun `should throw exception when api url is blank`() {
         MiniAppSdkConfig(
             baseUrl = " ",
-            isTestMode = true,
-            rasAppId = TEST_HA_ID_APP,
+            isPreviewMode = true,
+            rasProjectId = TEST_HA_ID_PROJECT,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
-            hostAppVersionId = TEST_HA_ID_VERSION,
             hostAppUserAgentInfo = TEST_HA_NAME
         )
     }
 
     @Test(expected = MiniAppSdkException::class)
-    fun `should throw exception when rasAppId is blank`() {
+    fun `should throw exception when projectId is blank`() {
         MiniAppSdkConfig(
             baseUrl = TEST_URL_HTTPS_2,
-            isTestMode = true,
-            rasAppId = " ",
+            isPreviewMode = true,
+            rasProjectId = " ",
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
-            hostAppVersionId = TEST_HA_ID_VERSION,
             hostAppUserAgentInfo = TEST_HA_NAME
         )
     }
@@ -59,22 +55,9 @@ class MiniAppSdkConfigSpec {
     fun `should throw exception when subscriptionKey is blank`() {
         MiniAppSdkConfig(
             baseUrl = TEST_URL_HTTPS_2,
-            isTestMode = true,
-            rasAppId = TEST_HA_ID_APP,
+            isPreviewMode = true,
+            rasProjectId = TEST_HA_ID_PROJECT,
             subscriptionKey = " ",
-            hostAppVersionId = TEST_HA_ID_VERSION,
-            hostAppUserAgentInfo = TEST_HA_NAME
-        )
-    }
-
-    @Test(expected = MiniAppSdkException::class)
-    fun `should throw exception when hostAppVersionId is blank`() {
-        MiniAppSdkConfig(
-            baseUrl = TEST_URL_HTTPS_2,
-            isTestMode = true,
-            rasAppId = TEST_HA_ID_APP,
-            subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
-            hostAppVersionId = " ",
             hostAppUserAgentInfo = TEST_HA_NAME
         )
     }