+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/RealMiniAppSpec.kt
@@ -122,9 +122,10 @@ class RealMiniAppSpec {
         val miniAppSdkConfig = MiniAppSdkConfig(
             baseUrl = TEST_URL_HTTPS_2,
             isTestMode = true,
-            rasAppId = TEST_HA_ID_APP,
+            rasProjectId = TEST_HA_ID_PROJECT,
+            isPreviewMode = true,
+            rasAppId = TEST_HA_ID_PROJECT,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
-            hostAppVersionId = TEST_HA_ID_VERSION,
             hostAppUserAgentInfo = TEST_HA_NAME
         )