+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/api/AppInfoApiSpec.kt
@@ -55,25 +55,12 @@ class AppInfoApiRequestSpec : AppInfoApiSpec() {
         val requestUrl = mockServer.takeRequest().requestUrl!!
 
         requestUrl.encodedPath shouldEndWith "miniapps/"
-        requestUrl.encodedQuery.toString() shouldContain "hostVersion=$TEST_HA_ID_VERSION"
-    }
-
-    @Test
-    fun `should fetch mini apps for the provided host app version in test mode`() {
-        mockServer.enqueue(createResponse())
-        retrofit.create(AppInfoApi::class.java)
-            .list(hostAppId = TEST_HA_ID_APP, hostAppVersionId = TEST_HA_ID_VERSION, testPath = "test")
-            .execute()
-        val request = mockServer.takeRequest()
-
-        request.requestUrl!!.encodedPath shouldEndWith "miniapps/test"
-        request.path!! shouldContain "test_version"
     }
 
     private fun executeListingCallByRetrofit() {
         mockServer.enqueue(createResponse())
         retrofit.create(AppInfoApi::class.java)
-            .list(hostAppId = TEST_HA_ID_APP, hostAppVersionId = TEST_HA_ID_VERSION)
+            .list(hostAppId = TEST_HA_ID_PROJECT)
             .execute()
     }
 }
@@ -87,7 +74,7 @@ class AppInfoApiResponseSpec : AppInfoApiSpec() {
         mockServer.enqueue(createResponse())
 
         miniAppInfo = retrofit.create(AppInfoApi::class.java)
-            .list(hostAppId = TEST_HA_ID_APP, hostAppVersionId = TEST_HA_ID_VERSION)
+            .list(hostAppId = TEST_HA_ID_PROJECT)
             .execute().body()!![0]
     }